feat: Cria metodo calcularValorTotal em ReservaService

+++ src/main/java/br/csi/Dormez/DTO/ReservaDTO.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -17,7 +18,7 @@ public class ReservaDTO {
 
     private LocalDateTime checkIn;
     private LocalDateTime checkOut;
-    private double valorTotal;
+    private BigDecimal valorTotal;
     private String status;
     private Long funcionarioId;
     private Long quartoId;
